requirements.txt
torch
transformers
pillow
accelerate
streamlit
